Move hooks folder into core and lantern
