+++ src/lantern/LanternLayout.js
@@ -24,8 +24,8 @@ import TopBar from "../components/TopBar/TopBar";
 import TopBarLogo from "../components/TopBar/TopBarLogo";
 import TopBarUserMenuForAuthenticatedUser from "../components/TopBar/TopBarUserMenuForAuthenticatedUser";
 import Footer from "../components/Footer";
-import usePageLayout from "../hooks/usePageLayout";
-import useIntercom from "../hooks/useIntercom";
+import usePageLayout from "./hooks/usePageLayout";
+import useIntercom from "./hooks/useIntercom";
 import { setTranslateLocale } from "../utils/i18nSettings";
 import { useRootStore } from "../components/StoreProvider";
