+++ src/lantern/RevocationCharts/RevocationCharts.js
@@ -29,7 +29,7 @@ import RevocationsByGender from "../RevocationsByGender";
 import RevocationsByRace from "../RevocationsByRace";
 import RevocationsByDistrict from "../RevocationsByDistrict";
 import { useRootStore } from "../../components/StoreProvider";
-import { useContainerHeight } from "../../hooks/useContainerHeight";
+import { useContainerHeight } from "../hooks/useContainerHeight";
 import "./RevocationCharts.scss";
 
 const CHARTS = [
