+++ src/lantern/hooks/__tests__/useIntercom.test.js
@@ -17,9 +17,9 @@
 
 import { renderHook } from "@testing-library/react-hooks";
 import useIntercom from "../useIntercom";
-import { useRootStore } from "../../components/StoreProvider";
+import { useRootStore } from "../../../components/StoreProvider";
 
-jest.mock("../../components/StoreProvider");
+jest.mock("../../../components/StoreProvider");
 
 describe("useIntercom hook tests", () => {
   const mockName = "some user name";
